Update loc reporter
